// src/api/model/categoryResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SilkwormAPI
 * API для интернет-магазина
 * OpenAPI spec version: v1
 */

export interface CategoryResponseDto {
  Id: string;
  /** @minLength 1 */
  Title: string;
  /** @minLength 1 */
  Slug: string;
  /** @nullable */
  ParentId?: string | null;
}