// src/api/model/createProductDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SilkwormAPI
 * API для интернет-магазина
 * OpenAPI spec version: v1
 */
import type { CreateProductVariantDto } from './createProductVariantDto';

export interface CreateProductDto {
  /**
   * @minLength 1
   * @maxLength 200
   */
  Title: string;
  /** @minLength 1 */
  Description: string;
  /** @minItems 1 */
  Images: string[];
  /**
   * @minLength 1
   * @maxLength 500
   */
  Preview: string;
  /**
   * @maxLength 500
   * @nullable
   */
  Decoration?: string | null;
  /** @minimum 0.01 */
  Price: number;
  /**
   * @minimum 0
   * @nullable
   */
  DiscountPrice?: number | null;
  CategoryId: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  Composition: string;
  /** @minItems 1 */
  Variants: CreateProductVariantDto[];
}