// src/api/model/productDetailsResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SilkwormAPI
 * API для интернет-магазина
 * OpenAPI spec version: v1
 */
import type { ProductVariantResponseDto } from './productVariantResponseDto';

export interface ProductDetailsResponseDto {
  Id: string;
  /** @minLength 1 */
  Title: string;
  /** @minLength 1 */
  Description: string;
  Images: string[];
  /** @minLength 1 */
  Preview: string;
  /** @nullable */
  Decoration?: string | null;
  Price: number;
  /** @nullable */
  DiscountPrice?: number | null;
  CategoryId: string;
  /** @minLength 1 */
  CategoryTitle: string;
  /** @minLength 1 */
  Composition: string;
  Variants: ProductVariantResponseDto[];
}