// src/api/model/productDetailsResponseDtoPaginatedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SilkwormAPI
 * API для интернет-магазина
 * OpenAPI spec version: v1
 */
import type { ProductDetailsResponseDto } from './productDetailsResponseDto';

export interface ProductDetailsResponseDtoPaginatedResponse {
  /** @nullable */
  Items?: ProductDetailsResponseDto[] | null;
  Total?: number;
  Page?: number;
  PageSize?: number;
  readonly TotalPages?: number;
}