// src/api/model/updateProductDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SilkwormAPI
 * API для интернет-магазина
 * OpenAPI spec version: v1
 */

export interface UpdateProductDto {
  /**
   * @maxLength 200
   * @nullable
   */
  Title?: string | null;
  /** @nullable */
  Description?: string | null;
  /**
   * @minItems 1
   * @nullable
   */
  Images?: string[] | null;
  /**
   * @maxLength 500
   * @nullable
   */
  Preview?: string | null;
  /**
   * @maxLength 500
   * @nullable
   */
  Decoration?: string | null;
  /**
   * @minimum 0.01
   * @nullable
   */
  Price?: number | null;
  /**
   * @minimum 0
   * @nullable
   */
  DiscountPrice?: number | null;
  /** @nullable */
  CategoryId?: string | null;
  /**
   * @maxLength 1000
   * @nullable
   */
  Composition?: string | null;
}